+++ src/Response/SpacesResponse.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Storyblok\ManagementApi\Response;
 
-use Storyblok\ManagementApi\Data\SpaceData;
-use Storyblok\ManagementApi\Data\SpacesData;
+use Storyblok\ManagementApi\Data\Space;
+use Storyblok\ManagementApi\Data\Spaces;
 use Storyblok\ManagementApi\Data\StoryblokData;
 use Storyblok\ManagementApi\Exceptions\StoryblokFormatException;
 use Storyblok\ManagementApi\Response\StoryblokResponse;
@@ -13,12 +13,12 @@ use Storyblok\ManagementApi\Response\StoryblokResponse;
 class SpacesResponse extends StoryblokResponse implements StoryblokResponseInterface
 {
     #[\Override]
-    public function data(): SpacesData
+    public function data(): Spaces
     {
         $key = "spaces";
         $array = $this->toArray();
         if (array_key_exists($key, $array)) {
-            return new SpacesData($array[$key]);
+            return new Spaces($array[$key]);
         }
 
         throw new StoryblokFormatException(sprintf("Expected '%s' in the response.", $key));